Split stepper test from components that require geant4
